+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Valkyrie - Changelog
 
+## v0.1.0 (2020-MAY-27)
+- supports method Cause
+- supports method Unwrap for Go 1.13 error chains
+
 ## v0.0.1 (2018-JAN-19)
 
 - initial fork commit